Synthesis of trees of gates – Fanin Tree Embedding
Introduction to Algorithms Course :

Nature of the project : A study and implementation of construction of
Fanin Trees to minimize the output delay and keeping the cost within the
given threshold. The solutions iterates through each of the possible
combination of gates that can made with the given inputs and retrieves
the optimal solution recursively. For n arrival times, the given
combinations can be n-1 ways. For instance, if the given arrival times
are 4 in number (ABCD), then there are 3 ways to form gates: A(BCD),
(AB)CD and (ABC)D. So we divide the problem into sub problems of smaller
size. Therefore, the problem has optimal substructure property and is
solved using recursion.

Tools used : C, Java, Visual Studio 2015.

Role and Responsibilities : Algorithm to construct a legal tree of gates
such that arrival time at the output of the tree is optimum,
implementing Recursion and Memoization to achieve the desired result.

Reward : We received extra credits for implementing a published IEEE
paper.

References : [1] Milos Hrkic, John Lillis, Giancarlo Beraudo title- “An
approach to Placement – Coupled Logic Replication” Topic – “Fannin Tree”

SKILLS